requirements.txt
APScheduler==3.10.4
SQLAlchemy>=2.0.0
python-dotenv==1.0.0